refactor: split dashboard summary into separate components
