+++ frontend/app/src/modules/dashboard/summary/ExchangeBox.vue
@@ -1,48 +1,46 @@
 <script setup lang="ts">
 import { Routes } from '@/router/routes';
-import { useGeneralSettingsStore } from '@/store/settings/general';
 import { useLocations } from '@/composables/locations';
 import AmountDisplay from '@/components/display/amount/AmountDisplay.vue';
 import LocationDisplay from '@/components/history/LocationDisplay.vue';
 import ListItem from '@/components/common/ListItem.vue';
 import type { BigNumber } from '@rotki/common';
 
-const props = defineProps<{
-  name: string;
+interface ExchangeBoxProps {
+  location: string;
   amount: BigNumber;
-}>();
+}
 
-const { name } = toRefs(props);
+const props = defineProps<ExchangeBoxProps>();
 
-const { currencySymbol } = storeToRefs(useGeneralSettingsStore());
+const { exchangeName } = useLocations();
 
-const manualBalancesRoute = Routes.BALANCES_MANUAL;
-
-const { locationData } = useLocations();
-const location = locationData(name);
+const exchangeLocationRoute = computed<string>(() => {
+  const route = Routes.BALANCES_EXCHANGE;
+  return `${route}/${props.location}`;
+});
 </script>
 
 <template>
-  <RouterLink :to="manualBalancesRoute">
+  <RouterLink :to="exchangeLocationRoute">
     <ListItem
-      data-cy="manual-balance__summary"
-      class="group py-1 px-6"
-      :data-location="name"
+      :id="`${location}_box`"
+      class="exchange-box__item group py-1 px-6"
     >
       <template #avatar>
         <div class="grayscale group-hover:grayscale-0">
           <LocationDisplay
-            :identifier="name"
+            :identifier="location"
             icon
             size="26px"
           />
         </div>
       </template>
       <div class="flex flex-wrap justify-between gap-1 text-rui-text">
-        {{ location?.name || toCapitalCase(name) }}
+        {{ exchangeName(location) }}
         <AmountDisplay
           show-currency="symbol"
-          :fiat-currency="currencySymbol"
+          fiat-currency="USD"
           :value="amount"
           class="font-medium"
         />
