+++ frontend/app/src/modules/dashboard/summary/Summary.vue
@@ -1,0 +1,13 @@
+<script lang="ts" setup>
+import ExchangeSummary from './ExchangeSummary.vue';
+import BlockchainSummary from './BlockchainSummary.vue';
+import ManualSummary from './ManualSummary.vue';
+</script>
+
+<template>
+  <div class="grid grid-cols-1 lg:grid-cols-3 gap-6 w-full">
+    <ExchangeSummary />
+    <BlockchainSummary />
+    <ManualSummary />
+  </div>
+</template>
