+++ frontend/app/src/modules/dashboard/summary/SummaryCardCreateButton.vue
